% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jseq.R
\name{jseq}
\alias{jseq}
\title{J-Optimal Subsample Selection}
\usage{
jseq(
  x,
  alpha,
  model_vec,
  k0 = 5 * length(model_vec),
  q = 5/8,
  gamma = 1/10,
  eps1 = 0
)
}
\arguments{
\item{x}{A dataset (data frame) containing the covariates for the regression model.}

\item{alpha}{A numeric value between 0 and 1 specifying the subsample proportion.}

\item{model_vec}{A character vector defining the regression model. Each element should
represent a term in the model, written as an expression involving \code{x}. For example,
\code{"1"} for the intercept, \code{"x[1]"} for the first covariate, or \code{"x[1]*x[2]^2"}
for an interaction term.}

\item{k0}{An integer specifying the initial size of the subsample. Defaults to \code{5*length(model_vec)}.}

\item{q}{A numeric value between \code{0.5} and \code{1}. Defaults to \code{5/8}.}

\item{gamma}{A numeric value between \code{0} and \code{q-0.5}. Defaults to \code{1/10}.}

\item{eps1}{A small positive value. Defaults to \code{0}.}
}
\value{
A list with the following components:
\item{x_j}{A subsample of \code{x} containing the selected observations (rows) according to J-optimality.}
\item{idx}{A vector of indices corresponding to the selected rows of \code{x}.}
}
\description{
This function implements the J-optimal subsample selection method, as described in
Cia-Mina et al. (2025). It takes a dataset of covariates \code{x}, a subsample
proportion \code{alpha} (between 0 and 1), and a vector defining a regression model.
Additional parameters can be specified to control the selection process.
}
\details{
The J-optimal subsample selection algorithm selects a subset of observations from the dataset \code{x}
that optimizes the statistical efficiency of the model defined by \code{model_vec}. For technical details,
refer to Cia-Mina et al. (2025).
}
\examples{
# Example 1: Bivariate regression
set.seed(123)
x1 <- runif(1e3, min = -1, max = 1)
x2 <- runif(1e3, min = -1, max = 1)
x <- data.frame(x1 = x1, x2 = x2)
model_vec <- c("1", "x[1]", "x[2]", "x[1]*x[2]", "x[1]^2", "x[2]^2")
result <- jseq(x, 0.3, model_vec)

# Plot the full dataset and the selected subsample
plot(x$x1, x$x2, col = "black", pch = 16, cex = 0.7, xlab = "x1", ylab = "x2")
points(result$x_j$x1, result$x_j$x2, col = "red", pch = 16, cex = 0.7)
title(main = "J-OPT", line = 1)

# Example 2: Univariate regression
set.seed(123)
x <- data.frame(x = rnorm(1e4))
model_vec <- c("1", "x[1]", "x[1]^2")
result <- jseq(x, 0.3, model_vec)

# Plot the density of the selected subsample
plot(density(result$x_j$x, bw = 2 / 100, kernel = "epanechnikov"),
     ylab = "", lwd = 1.7, xlim = c(-3.5, 3.5), main = "", xlab = "")

}
