% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_efficiency_comparison_example.R
\name{run_efficiency_comparison_example}
\alias{run_efficiency_comparison_example}
\title{Run Efficiency Comparison Example}
\usage{
run_efficiency_comparison_example()
}
\description{
This function provides an example of running an efficiency comparison
using the package's capabilities. The function contains the example
code previously executed at the top level.
}
\examples{
# To run the example:
run_efficiency_comparison_example()

}
