% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_efficiency_comparison_example.R
\name{run_efficiency_comparison_example}
\alias{run_efficiency_comparison_example}
\title{Run Efficiency Comparison Example}
\usage{
run_efficiency_comparison_example()
}
\description{
This function provides an example of running an efficiency comparison.
It compares the efficiency of two subdata selection methods:
J-optimal and D-optimal. It evaluates their performance based on
predefined efficiency criteria and generates comparative plots
to visualize the results.
Theoretical J-optimal is included for comparison.
}
\examples{
# To run the example:
run_efficiency_comparison_example()

}
